moved BSP to it's own file
